fix: api postKeyword 위치 변경
