+++ src/api/user/hooks.ts
@@ -1,5 +1,5 @@
-import { useQuery } from '@tanstack/react-query';
-import { getInterestedKeywords, getMentionUsers } from '.';
+import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
+import { getInterestedKeywords, getMentionUsers, KeywordSettingOptionType, postInterestedKeywards } from '.';
 
 export const useQueryGetMentionUsers = () => {
   return useQuery({
@@ -13,5 +13,16 @@ export const useQueryGetInterestedKeywords = () => {
   return useQuery({
     queryKey: ['getInterestedKeywords'],
     queryFn: () => getInterestedKeywords(),
+  });
+};
+
+export const useMutationInterestedKeywords = () => {
+  const queryClient = useQueryClient();
+
+  return useMutation({
+    mutationFn: (keywords: KeywordSettingOptionType[]) => postInterestedKeywards(keywords),
+    onSuccess: () => {
+      queryClient.invalidateQueries(['getInterestedKeywords']);
+    },
   });
 };
